+++ protocols/ping/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 - Raise MSRV to 1.65.
   See [PR 3715].
+- Remove deprecated items. See [PR 3702].
 
 [PR 3715]: https://github.com/libp2p/rust-libp2p/pull/3715
+[PR 3702]: https://github.com/libp2p/rust-libp2p/pull/3702
 
 ## 0.42.0